cleaning up some final details
