+++ assets/src/scripts/monitoring-location/iv-data-utils.js
@@ -6,6 +6,66 @@ import config from 'ui/config';
 import {convertCelsiusToFahrenheit} from 'ui/utils';
 
 /*
+* Helper function that determines if URL query parameter for time period is
+* acceptable. The purpose here is to prevent users from altering the URL in a way that
+* is undesirable or impossible for the application to process. We will accept
+* time periods in the form of P{MAX_DIGITS_FOR_DAYS_FROM_TODAY}D, or P1Y
+* @param {String} periodCode the time period code from the timespan radio buttons
+* @return {Boolean} if the value is within acceptable limits
+* */
+export const isPeriodWithinAcceptableRange = function(periodCode) {
+
+    return periodCode &&
+        periodCode.charAt(0) === 'P' &&
+        periodCode.slice(-1) === 'D' &&
+        periodCode.slice(1,-1).length < config.MAX_DIGITS_FOR_DAYS_FROM_TODAY ||
+        periodCode === 'P1Y';
+};
+
+
+/*
+* Helper function that sorts between 'custom' (user defined)  and the default time period options
+* @param {String} periodCode the time period code from the timespan radio buttons
+* @return {Boolean} if the value is or is not a 'custom' (user defined) time period
+* */
+export const isPeriodCustom = function(periodCode) {
+
+    return periodCode !== 'P7D' &&
+        periodCode !== 'P30D' &&
+        periodCode !== 'P1Y';
+};
+
+
+/*
+* Helper function that sorts between 'custom' (user defined)  and the default time period options
+* There are two general categories of user selected time periods (noted as 'period' in the code) for display on the hydrograph.
+* The first category contains time periods that are pre-defined in the application like 7 days, 30 days, and one year.
+* The second category are 'custom' time periods defined by the user through use of a calender date picker or days in an input field.
+* The 'period' information is passed in the URL as a parameter (such as &period=P1D). In the following line the period is
+* parsed into a 'custom' (or not) category and then parts are used to set the checked radio buttons and form fields as needed
+* @param {String} the time period code from the timespan radio buttons
+* @return {Object} parsed period code values
+* */
+export const parsePeriodCode = function(periodCode) {
+    if (periodCode) {
+        const mainTimeRangeSelectionButton = isPeriodCustom(periodCode) ? 'custom' : periodCode;
+        const userInputNumberOfDays = mainTimeRangeSelectionButton === 'custom' ? periodCode.slice(1,-1) : '';
+
+        return {
+            'mainTimeRangeSelectionButton' : mainTimeRangeSelectionButton,
+            'numberOfDaysFieldValue': userInputNumberOfDays
+        };
+    } else {
+
+        return {
+            'mainTimeRangeSelectionButton' : 'P7D',
+            'numberOfDaysFieldValue': ''
+        };
+    }
+};
+
+
+/*
 * Avoids having two Fahrenheit parameters for monitoring locations that already have both a measured Celsius and
 * Fahrenheit value for corresponding data parameters.
 *  Below is a listing of the known codes and counterparts
@@ -20,19 +80,18 @@ import {convertCelsiusToFahrenheit} from 'ui/utils';
 const checkForMeasuredFahrenheitParameters = function(parameterCode, NWISVariables) {
     let isParameterMatching;
     const allVariableParameterCodes = Object.entries(NWISVariables).map(variable => variable[1].variableCode.value);
-    const celsiusTemperatureCodesWithFahrenheitCounterparts = ['00020', '00010', '45589'];
     // If the current parameter code is one that could match a measured Fahrenheit parameter, cross reference the
     // parameter with the known corresponding codes that may match a value in the application state.
-    if (celsiusTemperatureCodesWithFahrenheitCounterparts.includes(parameterCode)) {
+    if (config.CELSIUS_CODES_WITH_FAHRENHEIT_COUNTERPARTS.includes(parameterCode)) {
         switch(parameterCode) {
             case '00020':
-                allVariableParameterCodes.includes('00021')? isParameterMatching = true : isParameterMatching = false;
+                isParameterMatching = allVariableParameterCodes.includes('00021');
                 break;
             case '00010':
-                allVariableParameterCodes.includes('00011')? isParameterMatching = true : isParameterMatching = false;
+                isParameterMatching = allVariableParameterCodes.includes('00011');
                 break;
             case '45589':
-                allVariableParameterCodes.includes('45590')? isParameterMatching = true : isParameterMatching = false;
+                isParameterMatching = allVariableParameterCodes.includes('45590');
                 break;
             default:
                 isParameterMatching = false;
@@ -43,6 +102,7 @@ const checkForMeasuredFahrenheitParameters = function(parameterCode, NWISVariabl
 
     return isParameterMatching;
 };
+
 
 /*
 * Converts a Celsius 'variable' from the NWIS system to one we can use to show Fahrenheit.
@@ -74,7 +134,7 @@ const createConvertedVariable = function(NWISVariable) {
  */
 const createConvertedTimeSeries = function(timeSeries, tsRequestKey, parameterCode) {
     timeSeries.points.forEach(temperaturePoint => {
-        temperaturePoint.value = convertCelsiusToFahrenheit(temperaturePoint.value);
+        temperaturePoint.value = convertCelsiusToFahrenheit(temperaturePoint.value).toFixed(2);
     });
 
     timeSeries.variable = `${timeSeries.variable}${config.CALCULATED_TEMPERATURE_VARIABLE_CODE}`;
